+++ blockfrost-api/src/Blockfrost/Types/Cardano/Addresses.hs
@@ -8,10 +8,12 @@ module Blockfrost.Types.Cardano.Addresses
   , AddressTransaction (..)
   ) where
 
-import Blockfrost.Types.Shared
 import Deriving.Aeson
 import qualified Money
 import Servant.Docs (ToSample (..), samples, singleSample)
+
+import Blockfrost.Types.Cardano.Scripts (InlineDatum (..))
+import Blockfrost.Types.Shared
 
 -- | Information about Cardano address
 data AddressInfo = AddressInfo
@@ -84,6 +86,8 @@ data AddressUtxo = AddressUtxo
   , _addressUtxoAmount      :: [Amount] -- ^ Amounts of Lovelaces or tokens
   , _addressUtxoBlock       :: BlockHash -- ^ Block hash of the UTXO
   , _addressUtxoDataHash    :: Maybe DatumHash -- ^ The hash of the transaction output datum
+  , _addressUtxoInlineDatum :: Maybe InlineDatum -- ^ CBOR encoded inline datum
+  , _addressUtxoReferenceScriptHash :: Maybe ScriptHash -- ^ The hash of the reference script of the output
   } deriving stock (Show, Eq, Generic)
   deriving (FromJSON, ToJSON)
   via CustomJSON '[FieldLabelModifier '[StripPrefix "_addressUtxo", CamelToSnake]] AddressUtxo
@@ -96,6 +100,8 @@ instance ToSample AddressUtxo where
       , _addressUtxoAmount = [ AdaAmount 42000000 ]
       , _addressUtxoBlock = "7eb8e27d18686c7db9a18f8bbcfe34e3fed6e047afaa2d969904d15e934847e6"
       , _addressUtxoDataHash = Just "9e478573ab81ea7a8e31891ce0648b81229f408d596a3483e6f4f9b92d3cf710"
+      , _addressUtxoInlineDatum = Nothing
+      , _addressUtxoReferenceScriptHash = Nothing
       }
     , AddressUtxo
       { _addressUtxoTxHash = "4c4e67bafa15e742c13c592b65c8f74c769cd7d9af04c848099672d1ba391b49"
@@ -103,6 +109,8 @@ instance ToSample AddressUtxo where
       , _addressUtxoAmount = [ AdaAmount 729235000 ]
       , _addressUtxoBlock = "953f1b80eb7c11a7ffcd67cbd4fde66e824a451aca5a4065725e5174b81685b7"
       , _addressUtxoDataHash = Nothing
+      , _addressUtxoInlineDatum = Nothing
+      , _addressUtxoReferenceScriptHash = Nothing
       }
     , AddressUtxo
       { _addressUtxoTxHash = "768c63e27a1c816a83dc7b07e78af673b2400de8849ea7e7b734ae1333d100d2"
@@ -117,6 +125,8 @@ instance ToSample AddressUtxo where
           ]
       , _addressUtxoBlock = "5c571f83fe6c784d3fbc223792627ccf0eea96773100f9aedecf8b1eda4544d7"
       , _addressUtxoDataHash = Nothing
+      , _addressUtxoInlineDatum = Nothing
+      , _addressUtxoReferenceScriptHash = Nothing
       }
     ]